refactor: index.ts에 Footer 컴포넌트 추가
