+++ src/app/place/components/index.ts
@@ -1,3 +1,4 @@
+export { default as Footer } from './Footer';
 export { default as Header } from './Header';
 export { default as ReviewItem } from './ReviewItem';
 export { default as ReviewSheet } from './ReviewSheet';
